Fix (Hopefully) the music bug
The music would sometimes be aborted with a not so easy to define error.
Apparently, the error was related to the readStream created by ytdl-core, in which the
bytes of the stream weren't generated fast enough. To fix this, I increased the value of
`highWaterMark`, which means that now there is more bytes of the song stored in memory (around 32mb)
giving plenty of room to generate all needed bytes, even when computation is slow.